pyproject.toml
[tool.poetry]
name = "omni"
version = "0.0.1"
description = "Omni"
authors = ["Matt Krainski <mateusz@krainski.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12,<3.13"
apache-airflow = "^2.10.4"
notion-client = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
darglint = "^1.8.1"
flake8 = "^7.1.1"
flake8-bandit = "^4.1.1"
flake8-bugbear = "^24.10.31"
flake8-builtins = "^2.5.0"
flake8-comprehensions = "^3.16.0"
flake8-docstrings = "^1.7.0"
flake8-eradicate = "^1.5.0"
flake8-isort = "^6.1.1"
flake8-spellcheck = "^0.28.0"
pep8-naming = "^0.14.1"
flake8-pyproject = "^1.2.3"
black = "^24.10.0"
flake8-black = "^0.3.6"
isort = "^5.13.2"
coverage = "^7.6.9"
pre-commit = "^4.0.1"
ipython = "^8.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
dictionaries = ["en_US", "python", "technical"]
builtins-ignorelist = ["id"]
ignore = ["D104", "D100", "D101", "D106"]
per-file-ignores = """
    tests/**: S101
"""
whitelist = ".spellcheck-allowlist"
strictness = "long"

[tool.isort]
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
profile = "black"
known_first_party = []

[tool.poe]
envfile = ".env"

[tool.poe.tasks]
configure = "./scripts/configure.sh"
update = "./scripts/update.sh"
lint = "flake8 ."
test = 'coverage run -m pytest --showlocals'
hooks-run = "pre-commit run --all-files"
hooks-update = "pre-commit auto-update"
generate-secret-key = "python scripts/generate_secret.py"

[tool.poe.tasks.test-full]
sequence = [
  { ref = "test" }
]
